Changed interval from `u64` to `f64`

+++ smelter-api/src/output/mp4.rs
@@ -43,7 +43,7 @@ pub enum Mp4VideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval: Option<u64>,
+        keyframe_interval: Option<f64>,
 
         /// (**default=`"yuv420p"`**) Encoder pixel format.
         pixel_format: Option<PixelFormat>,
@@ -58,7 +58,7 @@ pub enum Mp4VideoEncoderOptions {
         bitrate: Option<VideoEncoderBitrate>,
 
         /// (**default=`5000`**) Interval between keyframes, in milliseconds.
-        keyframe_interval: Option<u64>,
+        keyframe_interval: Option<f64>,
     },
 }
